chat/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.23",
    "embedchain>=0.1.128",
    "gitpython>=3.1.45",
    "ollama>=0.5.3",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "streamlit-chat>=0.1.1",
    "youtube-transcript-api>=1.2.2",
]
